ragu_acceleration: clean up tests and reuse ragu testing machinery
